changed the routing structure according to post and get method.
Updated the router class and added the Post and Get method and also the direct method is accepting the type of request which is coming from the server.

The request class now contains a method function which determines what kind of request is coming from the server.

Updated the Router in index.php, entry point of the application.